# [SILOptimizer] Stop optimization after serialization if only emitting a module