This is an experiment to stop the compilation pipeline after we've serialized a SIL module. There are a bunch of optimizations that happen after serialization whose results are tossed completely.

There are some modest wins here for compiling the stdlib with `-emit-module -O` (measured on an iMac Pro)

| Config                                                                 | Time    |
|:-----------------------------------------------------------------------|:--------|
| Swift Standard Library                                                 | 53.411s |
| Swift Standard Library (from interface)                                | 42.264s |
| Swift Standard Library (Stopping Optimizations Early)                  | 41.129s |
| Swift Standard Library (Stopping Optimizations Early) (from interface) | 36.66s  |