@harlanhaskins 2018-12-04
This should not change the benchmarks at all.

@swift-ci please benchmark