@harlanhaskins 2018-12-05
LLDB failure is unrelated.

@swift-ci please test macOS platform