@harlanhaskins 2018-12-05
@swift-ci please smoke test